+++ src/keymap.ts
@@ -1,3 +1,4 @@
+import { editableOperation } from "./editableoperation";
 import { getMessage } from "./languages";
 import { Keymap, operation } from "./operation";
 
@@ -31,6 +32,7 @@ export const defaultKeymap: Keymap[] = [
     label: "move cursor to the beginning",
     description: getMessage("move_cursor_to_the_beginning"),
     operation: operation.moveToTOL,
+    editableOperation: editableOperation.move_cursor_to_the_beginning,
     ctrl: true,
     key: "a",
   },
@@ -39,6 +41,7 @@ export const defaultKeymap: Keymap[] = [
     label: "move cursor to the end",
     description: getMessage("move_cursor_to_the_end"),
     operation: operation.moveToEOL,
+    editableOperation: editableOperation.move_cursor_to_the_end,
     ctrl: true,
     key: "e",
   },
@@ -47,6 +50,7 @@ export const defaultKeymap: Keymap[] = [
     label: "delete to the end of the line",
     description: getMessage("delete_to_the_end_of_the_line"),
     operation: operation.deleteToEOL,
+    editableOperation: editableOperation.delete_to_the_end_of_the_line,
     ctrl: true,
     key: "k",
   },
@@ -55,6 +59,7 @@ export const defaultKeymap: Keymap[] = [
     label: "move cursor to the next character",
     description: getMessage("move_cursor_to_the_next_character"),
     operation: operation.moveToNextChar,
+    editableOperation: editableOperation.move_cursor_to_the_next_character,
     ctrl: true,
     key: "f",
   },
@@ -63,6 +68,7 @@ export const defaultKeymap: Keymap[] = [
     label: "move cursor to the previous character",
     description: getMessage("move_cursor_to_the_previous_character"),
     operation: operation.moveToPreviousChar,
+    editableOperation: editableOperation.move_cursor_to_the_previous_character,
     ctrl: true,
     key: "b",
   },
@@ -71,6 +77,7 @@ export const defaultKeymap: Keymap[] = [
     label: "move cursor to the next word",
     description: getMessage("move_cursor_to_the_next_word"),
     operation: operation.moveToEndOfWord,
+    editableOperation: editableOperation.move_cursor_to_the_next_word,
     alt: true,
     key: "f",
   },
@@ -79,6 +86,7 @@ export const defaultKeymap: Keymap[] = [
     label: "move cursor to the previous word",
     description: getMessage("move_cursor_to_the_previous_word"),
     operation: operation.moveToTopOfWord,
+    editableOperation: editableOperation.move_cursor_to_the_previous_word,
     alt: true,
     key: "b",
   },
@@ -87,6 +95,7 @@ export const defaultKeymap: Keymap[] = [
     label: "delete to the beginning of the line",
     description: getMessage("delete_to_the_beginning_of_the_line"),
     operation: operation.deleteToTOL,
+    editableOperation: editableOperation.delete_to_the_beginning_of_the_line,
     ctrl: true,
     key: "u",
   },